Tail php-server.log when a functional test fails Fixes #163
When a functional test fails, the only visible error details is a 500 returned from the server.
In such cases we need to display a more detailed error message to give an idea of what happened on the server.
It should be possible to show the last few lines from /tmp/php-server.log